+++ Dyna/docs/release-notes.md
@@ -5,6 +5,7 @@
 ## Upcoming release
 
 * Added custom card art for all cards.
+* The Burst Charge now counts as 3 damage against the Sport Orb.
 * The CAT duo now counts various other cards as Basic Shots.
 
 ## 1.0.6